Add lookups for individual identifier types
